src/python/requirements.txt
pydantic
pyyaml
elevenlabs